# Conflicts:
#	com_joomlaquiz_expert/admin/controllers/quizzes.php

+++ com_joomlaquiz_expert/admin/models/quizzes.php
@@ -171,7 +171,15 @@ class JoomlaquizModelQuizzes extends JModelList
 				$pool_conf_item->q_id = $new_quiz_id;
 				$database->insertObject('#__quiz_pool', $pool_conf_item);
 			}
-			$query = "SELECT c_id FROM #__quiz_t_question WHERE c_quiz_id = '".$quiz2copy['c_id']."'";
+
+            $query = "SELECT * FROM #__quiz_feed_option WHERE quiz_id = '".$quiz2copy['c_id']."'";
+            $feed = $database->SetQuery( $query )->loadObjectList();
+            foreach($feed as $feed_opt){
+                $feed_opt->quiz_id = $new_quiz_id;
+                $database->insertObject('#__quiz_feed_option', $feed_opt);
+            }
+
+            $query = "SELECT c_id FROM #__quiz_t_question WHERE c_quiz_id = '".$quiz2copy['c_id']."'";
 			$database->SetQuery( $query );
 			$cid = $database->loadColumn();
 			if (!is_array( $cid )) {
@@ -310,6 +318,25 @@ class JoomlaquizModelQuizzes extends JModelList
 		$db = JFactory::getDBO();
 		$option = "com_joomlaquiz";
 		if (count( $cid )) {
+			
+			$query = "DELETE FROM #__quiz_pool" . "\n WHERE " . $db->qn('q_id') . " IN ( ". implode(', ', $cid) . ")";
+            $db->setQuery( $query );
+            $db->execute();
+
+            $query = "DELETE FROM #__quiz_feed_option" . "\n WHERE " . $db->qn('quiz_id') . " IN ( ". implode(', ', $cid) . ")";
+            $db->setQuery( $query );
+            $db->execute();
+
+			$names = "'com_joomlaquiz.quiz.".implode( "', 'com_joomlaquiz.quiz.", $cid )."'";
+            $query = "DELETE FROM #__assets"
+                . "\n WHERE ".$db->qn('name')." IN ( $names )"
+            ;
+            $db->setQuery( $query );
+            if (!$db->execute()) {
+                echo "<script> alert('".$db->getErrorMsg()."'); window.history.go(-1); </script>\n";
+                exit();
+            }
+
 			$cids = implode( ',', $cid );
 			$query = "DELETE FROM #__quiz_t_quiz"
 			. "\n WHERE c_id IN ( $cids )"
